[Application] Move Directory class from Support namespace.
