+++ src/Valkyrja/Application/Directory/Directory.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Valkyrja\Support\Directory;
+namespace Valkyrja\Application\Directory;
 
 class Directory
 {
