+++ src/Valkyrja/Log/Provider/ServiceProvider.php
@@ -18,6 +18,7 @@ use Monolog\Handler\StreamHandler;
 use Monolog\Logger;
 use Override;
 use Psr\Log\LoggerInterface;
+use Valkyrja\Application\Directory\Directory;
 use Valkyrja\Application\Env\Env;
 use Valkyrja\Container\Manager\Contract\ContainerContract;
 use Valkyrja\Container\Provider\Provider;
@@ -25,7 +26,6 @@ use Valkyrja\Log\Enum\LogLevel;
 use Valkyrja\Log\Logger\Contract\LoggerContract;
 use Valkyrja\Log\Logger\NullLogger;
 use Valkyrja\Log\Logger\PsrLogger;
-use Valkyrja\Support\Directory\Directory;
 
 use function date;
