+++ src/Valkyrja/View/Provider/ServiceProvider.php
@@ -18,11 +18,11 @@ use Twig\Environment;
 use Twig\Error\LoaderError;
 use Twig\Extension\ExtensionInterface;
 use Twig\Loader\FilesystemLoader;
+use Valkyrja\Application\Directory\Directory;
 use Valkyrja\Application\Env\Env;
 use Valkyrja\Container\Manager\Contract\ContainerContract;
 use Valkyrja\Container\Provider\Provider;
 use Valkyrja\Http\Message\Response\Factory\Contract\ResponseFactoryContract as HttpMessageResponseFactory;
-use Valkyrja\Support\Directory\Directory;
 use Valkyrja\View\Factory\Contract\ResponseFactoryContract;
 use Valkyrja\View\Factory\ResponseFactory;
 use Valkyrja\View\Orka\Constant\OrkaReplacement;
