+++ tests/Tests/Functional/Abstract/TestCase.php
@@ -13,11 +13,11 @@ declare(strict_types=1);
 
 namespace Valkyrja\Tests\Functional\Abstract;
 
+use Valkyrja\Application\Directory\Directory;
 use Valkyrja\Application\Entry\Abstract\App;
 use Valkyrja\Application\Kernel\Contract\ApplicationContract;
 use Valkyrja\Http\Message\Request\Contract\ServerRequestContract;
 use Valkyrja\Http\Message\Request\Factory\RequestFactory;
-use Valkyrja\Support\Directory\Directory;
 use Valkyrja\Tests\Abstract\TestCase as AbstractTestCase;
 use Valkyrja\Tests\EnvClass;
